Remove polynomial commitment interface.
