+++ src/core/poly/mod.rs
@@ -1,4 +1,3 @@
 pub mod circle;
-pub mod commitment;
 pub mod line;
 mod utils;
