+++ src/fibonacci.rs
@@ -1,6 +1,8 @@
 use num_traits::One;
 
 use crate::commitment_scheme::hasher::Hasher;
+use crate::commitment_scheme::merkle_decommitment::MerkleDecommitment;
+use crate::commitment_scheme::merkle_tree::MerkleTree;
 use crate::core::air::{Mask, MaskItem};
 use crate::core::channel::{Blake2sChannel, Channel as ChannelTrait};
 use crate::core::circle::{CirclePoint, Coset};
@@ -12,8 +14,8 @@ use crate::core::fields::qm31::QM31;
 use crate::core::fields::{ExtensionOf, Field, IntoSlice};
 use crate::core::oods::{get_oods_quotient, get_oods_values};
 use crate::core::poly::circle::{CanonicCoset, CircleDomain, CircleEvaluation, PointMapping};
-use crate::core::poly::commitment::{CommitmentProof, PolynomialCommitmentScheme};
 use crate::core::queries::Queries;
+use crate::core::utils::bit_reverse_vec;
 
 type Channel = Blake2sChannel;
 type MerkleHasher = <Channel as ChannelTrait>::ChannelHasher;
@@ -35,6 +37,13 @@ pub struct AdditionalProofData {
     pub composition_polynomial_oods_value: QM31,
     pub composition_polynomial_random_coeff: QM31,
     pub oods_point: CirclePoint<QM31>,
+}
+
+// TODO(AlonH): Removed this struct and separate the decommitment from the the commitment in the
+// stark proof.
+pub struct CommitmentProof<F: ExtensionOf<BaseField>, H: Hasher> {
+    pub decommitment: MerkleDecommitment<F, H>,
+    pub commitment: H::Hash,
 }
 
 pub struct FibonacciProof {
@@ -181,9 +190,11 @@ impl Fibonacci {
         let trace_evaluation = trace_poly.evaluate(self.trace_eval_domain);
         let trace_commitment_evaluation =
             trace_poly.evaluate(self.trace_commitment_domain.circle_domain());
-        let trace_commitment = PolynomialCommitmentScheme::<BaseField, MerkleHasher>::commit(vec![
-            &trace_commitment_evaluation,
-        ]);
+        let trace_commitment =
+            MerkleTree::<BaseField, MerkleHasher>::commit(vec![bit_reverse_vec(
+                &trace_commitment_evaluation.values,
+                trace_commitment_evaluation.domain.log_size(),
+            )]);
         channel.mix_with_seed(trace_commitment.root());
 
         let random_coeff = channel.draw_random_extension_felts()[0];
@@ -195,9 +206,12 @@ impl Fibonacci {
                 .circle_domain(),
         );
         let composition_polynomial_commitment =
-            PolynomialCommitmentScheme::<QM31, MerkleHasher>::commit(vec![
-                &composition_polynomial_commitment_evaluation,
-            ]);
+            MerkleTree::<QM31, MerkleHasher>::commit(vec![bit_reverse_vec(
+                &composition_polynomial_commitment_evaluation.values,
+                composition_polynomial_commitment_evaluation
+                    .domain
+                    .log_size(),
+            )]);
         channel.mix_with_seed(composition_polynomial_commitment.root());
 
         let oods_point = CirclePoint::<QM31>::get_random_point(channel);
@@ -241,13 +255,14 @@ impl Fibonacci {
             .collect();
         // TODO(AlonH): Get sub circle domains from FRI.
         const FRI_STEP_SIZE: u32 = 1;
-        let composition_polynomial_decommitment = composition_polynomial_commitment.decommit(
-            &composition_polynomial_queries
-                .to_sparse_sub_circle_domain(FRI_STEP_SIZE)
-                .to_decommitment_positions(),
-        );
-        let trace_decommitment = trace_commitment.decommit(
-            &trace_queries
+        let composition_polynomial_decommitment = composition_polynomial_commitment
+            .generate_decommitment(
+                composition_polynomial_queries
+                    .to_sparse_sub_circle_domain(FRI_STEP_SIZE)
+                    .to_decommitment_positions(),
+            );
+        let trace_decommitment = trace_commitment.generate_decommitment(
+            trace_queries
                 .to_sparse_sub_circle_domain(FRI_STEP_SIZE)
                 .to_decommitment_positions(),
         );
